Solve #1956.
Diagnostics from gnu_time is now written to 2 places, sessiondir/.diag
and sessiondir/*.comment. The latter is transferred to the head node
via an LRMS-specific mechanism, as opposed to *.diag which relies on a
shared filesystem.

In the *.comment file, diagnostics data lines are prefixed with _DIAG_
to differentiate them from other content like error messages from RTE
scripts and messages from LRMS.

The SGE backend is modified to take advantage of the changes to the
*.comment file.