Update .gitlab-ci.yml - Removed the host certificate generation. And run deploy on separate machines for el6 and el7.
The host certificate generation is not needed as the runner machine is persistent, and the /etc folder is mounted inside the docker container as a volume. So no need to regenerate it each time. If however /etc was not mounted inside the docker container, then host-generation would have to be done for each docker container. 

Running deploy on same machine for el6 and el7 is a problem since arex start up scripts can be found in /etc which is available for all docker images so el6 deploy and el7 deploy will interfere. Are therefore now running the two deploys on separate machines.