Patch for compatibility with old ARC infoprovider modules
This patch adds proper logging support for Condor.pm and other
PERL modules who where written for arc 0.6.

It also includes new information in the datastructure that is passed to
the LRMS modules, namely loglevel and where the controldir is (a feature needed
by Condor.pm), as some of this got lost with the new configuration.