AC_CHECK_HEADER checks for the presence of the header (using the precompiler) and for the usability of the header (using the compiler). If the two tests disagree newer autotools versions use the result from the usability test.
However, older autotools versions (which include the version used by
the nightlies) use the presence test in this situation, which screws
up the logic of the xrootd test. This change reimplements the test
using AC_TRY_COMPILE so that the compile test is used also when older
autotools versions are used.