Fix reading of pickled config
Do not use "from .config import Config" since this will import the
current Config object at the time of import. If the Config object in
the .config module is replace later, this will then not change the
imported object.

Instead use "from . import config" and refer to the Config object as
config.Config. This way a redefinition of Config object in the .config
module works as expected.

This is important when using the pickled config, since the import of
the pickled config replaces the Config object.