Don't use Python asserts for program logic
I Python asserts are used the check that things that "should not
happen" do not happen. They are used to catch failures during
development and testing.

$ python3 -c 'assert(False); print("We are here")'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
AssertionError

But since what they test "should not happen", when running in
optimized mode, these asserts are optimized away:

$ python3 -O -c 'assert(False); print("We are here")'
We are here

So asserts can not be used for tests that are essential to the program
logic and should always be run.