Fix an infinite loop when the jids array is empty
The jobs_info routine matches the known jobs with those in
the batch system, but commit 91d2e601f introduces a bug
where the index in the jids array assumes that there is at
least one job.

Since the routine should return an empty hash in case jids
is empty, get this done as soon as possible before doing
any other work.