Restore ServiceEndpointRetrieverTest.py test for swig-4.2.1
swig-4.2.0 is a bit stricter for requiring correct type information
across imported modules. This patch allows compute.i to correctly import
the type information from common.i for the SimpleConditionList template
instantiation:

  %template(SimpleConditionList) std::list<Arc::SimpleCondition*>;

The type information is required when this template is used in TestACCControl.h:

  class ServiceEndpointRetrieverPluginTESTControl {
  public:
    static std::list<SimpleCondition*> condition;

and is now correctly available when wrapping ServiceEndpointRetrieverPluginTESTControl.

However, this requires swig-4.2.1 which corrects the way static member
variables are accessed, such as ServiceEndpointRetrieverPluginTESTControl::condition,
by creating wrappers that access the variable by pointer rather than by
value. In essence this means that when reading the condition variable it
is done via a pointer to std::list rather than by copying the entire contents
into a Python tuple.

Without this change, the wrapper is:
  SWIGINTERN PyObject *Swig_var__ServiceEndpointRetrieverPluginTESTControl_condition_get(void) {
    PyObject *pyobj = 0;

    pyobj = SWIG_NewPointerObj(SWIG_as_voidptr(&Arc::ServiceEndpointRetrieverPluginTESTControl::condition), SWIGTYPE_p_std__listT_Arc__SimpleCondition_p_std__allocatorT_Arc__SimpleCondition_p_t_t,  0 );
    return pyobj;
  }

where SWIGTYPE_p_std__listT_Arc__SimpleCondition_p_std__allocatorT_Arc__SimpleCondition_p_t_t
is the descriptor to an opaque pointer (unknown type) and this descriptor in computer_wrap.cpp
does not match that in common_wrap.cpp.

With swig-4.2.1 and this patch the wrapper is:

  SWIGINTERN PyObject *Swig_var__ServiceEndpointRetrieverPluginTESTControl_condition_get(void) {
    PyObject *pyobj = 0;

    pyobj = SWIG_NewPointerObj(SWIG_as_voidptr(&Arc::ServiceEndpointRetrieverPluginTESTControl::condition), SWIGTYPE_p_std__listT_Arc__SimpleCondition_p_t,  0 );
    return pyobj;
  }

and the SWIGTYPE_p_std__listT_Arc__SimpleCondition_p_t descriptor
in compute_wrap.cpp correctly matches that in common_wrap.cpp.

With swig-4.2.0 and this patch the wrapper is:

  SWIGINTERN PyObject *Swig_var__ServiceEndpointRetrieverPluginTESTControl_condition_get(void) {
    PyObject *pyobj = 0;

    pyobj = swig::from(static_cast< std::list< Arc::SimpleCondition*,std::allocator< Arc::SimpleCondition * > > >(Arc::ServiceEndpointRetrieverPluginTESTControl::condition));
    return pyobj;
  }

which is the code to create and return a tuple, that is, return a copy
of the static member variable by value. A Python tuple does not have a
'clear' method. swig-4.2.1 and this patch is required for this code in
ServiceEndpointRetrieverTest.py to work:

  ServiceEndpointRetrieverPluginTESTControl.condition.clear()


(cherry picked from commit f258468c5778c943079b95c71a25536c9469ae08)