Restore support for swig-4.2.0 and earlier
for ServiceEndpointerRetrieverPluginTestControl::condition wrappers

This clears the problematic typemap that is fixed in swig-4.2.1. The
cleared typemap was incorrectly returning the list as a python tuple by
value instead of wrapping as a pointer to a std::list.


(cherry picked from commit 31b55f1f5655ab7b50bbb3dd671d17fdce7067f3)