Reintroduce comparison between ARC-CE and Condor job status's
With commit c7a67687, the logic to compare job state with `condor_q` and job state in `/var/spool/arc/jobstatus/processing` was lost. This caused every job to be considered finished when the script `/usr/share/arc/scan-condor-job` ran.

This regression is corrected in this commit by reintroducing the `lrms_list_jobs` function and having it run when the function `scan_init` is called. The `scan_init` is purposefully removed from the previous for-loop as not to overwhelm busy Condor schedd's with multiple `condor_q` requests

During the `scan_main` function call, a comparison is done where the `condor_q` list is compared with what ARC see's in `/var/spool/arc/jobstatus/processing`, if there isn't a match, the job has finished in Condor and the script can continue, if not, exit the script cleanly.